+++ Examples/NativeTemplatesExample/NativeTemplatesExample/ViewController.m
@@ -24,11 +24,16 @@
 {
     [super viewDidLoad];
 
-    // Replace demo R-M-DEMO-native-c with actual Block ID
+    // Replace demo R-M-DEMO-native-c with actual Block ID.
+    // Please, note, that configured image sizes don't affect demo ads.
     // Following demo Block IDs may be used for testing:
     // R-M-DEMO-native-c
     // R-M-DEMO-native-i
-    self.adLoader = [[YMANativeAdLoader alloc] initWithBlockID:@"R-M-DEMO-native-c"];
+    YMANativeAdLoaderConfiguration *configuration =
+        [[YMANativeAdLoaderConfiguration alloc] initWithBlockID:@"R-M-DEMO-native-c"
+                                                     imageSizes:@[ kYMANativeImageSizeMedium ]
+                                        loadImagesAutomatically:YES];
+    self.adLoader = [[YMANativeAdLoader alloc] initWithConfiguration:configuration];
     self.adLoader.delegate = self;
     [self.adLoader loadAdWithRequest:nil];
 }